// src/core/api/generated/spring/schemas/UserAvailabilityDTODayOfWeek.ts
/**
 * Generated by orval v7.8.0 🍺
 * Do not edit manually.
 * Crm Backend API
 * Crm Backend API documentation
 * OpenAPI spec version: 0.0.1
 */

export type UserAvailabilityDTODayOfWeek = typeof UserAvailabilityDTODayOfWeek[keyof typeof UserAvailabilityDTODayOfWeek];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const UserAvailabilityDTODayOfWeek = {
  MONDAY: 'MONDAY',
  TUESDAY: 'TUESDAY',
  WEDNESDAY: 'WEDNESDAY',
  THURSDAY: 'THURSDAY',
  FRIDAY: 'FRIDAY',
  SATURDAY: 'SATURDAY',
  SUNDAY: 'SUNDAY',
} as const;